+++ CHANGELOG.md
@@ -6,13 +6,13 @@
 + [x] Server: added alert to DISCORD (used **discord webhooks**)
 + [x] Server: added config props `alertToTelegram`, `alertToDiscord`. Possible values: `["NOT-SYNCED", "MAX", "FORK", "FORWARD-FORK", "HANG", "EXPLORER", "RESTART", "BALANCE"]`
 + [x] Server: moved request to `coingecko` from client to server
++ [x] Server: added `config.price` parameter
 
 + [x] Client: added flashing for `maxHeight`, `unvalidateHeight`, `explorerHeight` when difference present with the current height
 + [x] Client: high/low price now showing **all-time** high/low  
 + [x] Client: added price change in currency, added colored arrow for indicate price change  
 + [x] Client: next `config` props deprecated and not used: `cpu`, `net`, `mem`, `blockchain`, `node`, `info`, `time`
 + [x] Client: added new config props: `system` (server info and time), `resources` (cpu, net, ram), `daemon` (data from Mina GraphQL: node info, blockchain status)
-+ [x] Client: renamed `config.price` to `config.coingecko`  
 + [x] Client: changed CPU temperature view
 + [x] Client: changed block height view
 + [x] Client: added border radius to panels